Refactored TazColor to SetColor in Const which implements a UserInterface depended Set of Colors
